+++ core/execution/market.go
@@ -2585,7 +2585,9 @@ func (m *Market) amendOrder(
 
 		// Update the existing message in place before we cancel it
 		if foundOnBook {
-			m.orderAmendInPlace(existingOrder, amendedOrder)
+			// Do not amend in place, the amend could be something
+			// not supported for an amend in place, and not pass
+			// the validation of the order book
 			cancellation, err := m.matching.CancelOrder(amendedOrder)
 			if cancellation == nil || err != nil {
 				m.log.Panic("Failure to cancel order from matching engine",
@@ -2593,11 +2595,10 @@ func (m *Market) amendOrder(
 					logging.String("order-id", amendedOrder.ID),
 					logging.String("market", m.mkt.ID),
 					logging.Error(err))
-				return nil, nil, err
 			}
 
-			_ = m.position.UnregisterOrder(ctx, cancellation.Order)
-			amendedOrder = cancellation.Order
+			// unregister the existing order
+			_ = m.position.UnregisterOrder(ctx, existingOrder)
 		}
 
 		// Update the order in our stores (will be marked as cancelled)